Split the codebase into the backend/frontend stages
