+++ src/frontend/compiler/scope.rs
